Three column display and selection of dirs.
Added linux/limits.h for PATH_MAX and sys/stat.h for determining if a
dirent is a directory.

The program can now handle basic movement using the hjkl keys. quit with
q key, and print current directory with <enter> key. To do this, the
path is associated with dir list, as opposed to each dir_entry.

The child directory is only printed if the highlighted file is actually
a directory.

The parent directory is only printed if the current is not root.
The directory selected is now highlighted.